Fix the NullAsConst canonicalizer rule
